Fixed the case when we detect an error while creating a new config.
In this case we have a back-up, but no new file.

+++ kicad_auto/file_util.py
@@ -123,7 +123,8 @@ def check_lib_table(fuser, fsys):
 
 def restore_one_config(name, fname, fbkp):
     if fbkp and os.path.exists(fbkp):
-        os.remove(fname)
+        if os.path.exists(fname):
+            os.remove(fname)
         os.rename(fbkp, fname)
         logger.debug('Restoring old %s config', name)
         return None
